Implemented suggestions from BlueRaja
Allow custom loggers to implement a custom log line formatter
